% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze_citations.R
\name{analyze_citations}
\alias{analyze_citations}
\title{Generates figures with summary of analysis of citation cases}
\usage{
analyze_citations(file, article, output)
}
\arguments{
\item{file}{'citation_data.csv' that contains data on the citation cases.}

\item{article}{Name of cited article; e.g. Fearon (2003)}

\item{output}{folder where figures generated by function will be stored}
}
\description{
Generates figures with summary of analysis of citation cases
}
\examples{
\dontrun{
 file <- "~/Google Drive/2016_Quality_of_citations/data/Fearon 2003_citation_cases.csv"
 article <- "Fearon and Laitin (2003)"
 output <- "fearon_2003"
 analyze_citations(file, article, output)
}
}

