% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_citations_cases.R
\name{clean_citation_cases}
\alias{clean_citation_cases}
\title{Takes the rather "roughly" identifed citation cases and cleans them.}
\usage{
clean_citation_cases(file)
}
\arguments{
\item{file}{'citation_data.csv' that contains data on the citation cases.}
}
\value{
Returns and saves cleaned version of citation cases data in 'citation_data_cleaned.csv'.
}
\description{
Takes the rather "roughly" identifed citation cases and cleans them.
}

