% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pdf_filenames.R
\name{clean_filenames}
\alias{clean_filenames}
\title{Clean text files in /documents/ folder for subsequent analysis.}
\usage{
clean_filenames()
}
\value{
Converts PDFs to text and stores in documents folder.
}
\description{
Clean text files in /documents/ folder for subsequent analysis.
}

