% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean text files in /documents/ folder for subsequent analysis.}
\usage{
clean_text(folder, number = NULL)
}
\arguments{
\item{folder}{The folder in which the PDFs are located.}
}
\value{
Converts PDFs to text and stores in documents folder.
}
\description{
Clean text files in /documents/ folder for subsequent analysis.
}

