% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_wos_citations.R
\name{count_wos_citations}
\alias{count_wos_citations}
\title{Count citations of study (or studies) in the web of knowledge.}
\usage{
count_wos_citations(wos.url)
}
\arguments{
\item{wos.url}{See steps below to get the respective url.}
}
\value{
Returns the number of times a study has been cited in the web of knowledge.
#'
}
\description{
Count citations of study (or studies) in the web of knowledge.
}
\section{Steps to follow}{

1. Search for study in WOS.

2. Click on times cited for the particular study you are interested in.

3. Click on the first result in the list.

4. Copy the link out of the adress line in the browser and insert that into the function.
}

