% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_ref_section.R
\name{delete_ref_section}
\alias{delete_ref_section}
\title{Returns and saves text files without the references section.}
\usage{
delete_ref_section(folder, number = NULL)
}
\arguments{
\item{folder}{Name of folder within working directory in which the citing documents (.txt files) are located, e.g. "Beck 1995".}

\item{number}{Number of .txt files in folder the function should be applied to. Default is "all .txt files in folder".}
}
\value{
Returns and saves text files without the references section.
}
\description{
Returns and saves text files without the references section.
}

