% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delete_reference_section.R
\name{delete_reference_section}
\alias{delete_reference_section}
\title{Deletes the references section in "_cleaned.txt" files and generates "cleaned_woref.txt" files.}
\usage{
delete_reference_section()
}
\value{
Returns and saves text files that do not contain a reference section anymore.
}
\description{
Deletes the references section in "_cleaned.txt" files and generates "cleaned_woref.txt" files.
}

