% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_refs_n_heads.R
\name{delete_refs_n_heads}
\alias{delete_refs_n_heads}
\title{Returns and saves text files without the references section.}
\usage{
delete_refs_n_heads(metadata = NULL, folder, number = NULL,
  encoding = "ASCII")
}
\arguments{
\item{folder}{Name of folder within working directory in which the citing documents (.txt files) are located, e.g. "Beck 1995".}

\item{number}{Number of .txt files in folder the function should be applied to. Default is "all .txt files in folder".}
}
\value{
Returns and saves text files without the references section.
}
\description{
Returns and saves text files without the references section.
}
\examples{
\dontrun{
 setwd("C:/Users/paul/Google Drive/Research/2016_Quality_of_citations/data")
 folder <- "Acemoglu 2001"
 delete_refs_n_heads(folder, number = 100)
}
}
