% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delete_running_heads.R
\name{delete_running_heads}
\alias{delete_running_heads}
\title{Deletes the references section in "_cleaned.txt" files and generates "cleaned_woref.txt" files.}
\usage{
delete_running_heads(folder, number = NULL)
}
\value{
Uses file name to try to identify if there are any running heads that
can be omitted from the document.
}
\description{
Deletes the references section in "_cleaned.txt" files and generates "cleaned_woref.txt" files.
}

