% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_names_encoding.R
\name{docs_names_encoding}
\alias{docs_names_encoding}
\title{Identifies meta information within txt documents.}
\usage{
docs_names_encoding(folder = "docs")
}
\arguments{
\item{folder}{Name of folder that contains the documents.}

\item{pdf}{Default TRUE. Set FALSE if only .txt files should be renamed.}

\item{txt}{Default TRUE. Set FALSE if only .pdf files should be renamed.}
}
\description{
Renames files in folder with rising index.
}
\examples{
\dontrun{
 setwd("C:/data")
 rename_docs(folder = "docs")
}
}
