% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_citation_cases.R
\name{extract_citation_cases}
\alias{extract_citation_cases}
\title{Extract citation cases.}
\usage{
extract_citation_cases(folder, authorname, studyyear, scope = NULL,
  number = NULL, clean = NULL)
}
\arguments{
\item{folder}{Name of folder within working directory in which the citing documents (.txt files) are located, e.g. "Beck 1995".}

\item{authorname}{Enter the names of all authors separated by comma.}

\item{studyyear}{Enter the year when the study was published.}

\item{scope}{Enter number of sentences before/after citation case sentence to extract.}

\item{clean}{If TRUE (default) it cleanes some stuff before the citation case}
}
\value{
Returns and saves data (.csv and .html) containing citation cases in current working directory.
}
\description{
Extract citation cases.
}
\examples{
\dontrun{
 setwd("C:/Users/paul/Google Drive/Research/2016_Quality_of_citations/data")
 folder <- "docs"
 study.title <- "The colonial origins of comparative development: An empirical investigation"
 authorname <- "Acemoglu, Johnson, Robinson"
 studyyear <- 2001
 extract_citation_cases(folder, authorname, studyyear, clean = TRUE) #
}
}

