% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{Converts and stores PDFs in folder as .txt files.}
\usage{
extract_text(folder, number = NULL)
}
\arguments{
\item{folder}{Name of folder within working directory in which the citing documents (PDFs) are located, e.g. "Beck 1995".}

\item{number}{Number of PDFs you want to apply the function to. Default is "all PDFs in folder".}
}
\value{
Converts and stores PDFs in folder as .txt files.
}
\description{
Converts and stores PDFs in folder as .txt files.
}

