% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch_paperpiledocs.R
\name{fetch_paperpiledocs}
\alias{fetch_paperpiledocs}
\title{Fetches documents from paperpile gdrive folder and copies to folder.}
\usage{
fetch_paperpiledocs(paperpile.directory)
}
\arguments{
\item{paperpile.directory}{The directory of the paperpile folder.}
}
\value{
Returns a .ris file that can be imported into paperpile.
}
\description{
Fetches documents from paperpile gdrive folder and copies to folder.
}

