% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identify_citing_document.R
\docType{data}
\name{folder}
\alias{folder}
\title{Identifies citation information in documents.}
\format{\preformatted{ chr "identify_citation_document"
}}
\usage{
folder
}
\arguments{
\item{folder}{The folder in which the citing documents are located.}
}
\value{
Returns a both a .csv file as well as a .html file that contain a
list with the titles that were extracted from all the citing documents.
}
\description{
Identifies citation information in documents.
}
\keyword{datasets}

