% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_ris.R
\name{gen_ris}
\alias{gen_ris}
\title{Generate .ris for all folders/studies in working directory and save in respective folders called ris.}
\usage{
gen_ris()
}
\value{
Returns a .ris file that can be imported into paperpile.
}
\description{
Generate .ris for all folders/studies in working directory and save in respective folders called ris.
}

