% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Identifies meta information within txt documents.}
\usage{
get_metadata(from, number = NULL, encoding = "ASCII", lines.import = 200,
  rcrossref = TRUE, bibtex = FALSE, vars = NULL, file = NULL,
  rename.file = FALSE, start = NULL, end = NULL)
}
\arguments{
\item{encoding}{Character encoding scheme of input txt file. Default is "utf-8".}

\item{lines.import}{Number of lines that should be imported from the input txt file. Default is 2000.}

\item{rcrossref}{Use rcrossref function to identify article meta data via CrossRef API? Default is TRUE. Currently, no useful other ways of identifying metadata are implemented.}

\item{bibtex}{Should reference be exported as bibtex? Works only with rcrossref = TRUE. Default is FALSE.}

\item{vars}{Which variables should be returned? Specify as character vectors. By default, all available variables are returned.}

\item{filename}{Input file; assumed to be txt format..}

\item{rename.files}{Rename the text files with information found online. Default = FALSE.}
}
\description{
Takes a single txt file as input. To operate on a vector of txt files, use get_metadata_doc().
}
\examples{
\dontrun{
 setwd("C:/data")
 number <- 20 # or do not specify
 get_metadata_doc_nv(folder)
 ...set also other arguments in the funcion....
}
}
