% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_doc.R
\name{get_metadata_doc}
\alias{get_metadata_doc}
\title{Identifies meta information within txt documents.}
\usage{
get_metadata_doc(filename, encoding = "UTF-8", lines.import = 2000,
  rcrossref = TRUE, bibtex = FALSE, vars = NULL)
}
\arguments{
\item{filename}{Input file; assumed to be txt format..}

\item{encoding}{Character encoding scheme of input txt file. Default is "utf-8".}

\item{lines.import}{Number of lines that should be imported from the input txt file. Default is 2000.}

\item{rcrossref}{Use rcrossref function to identify article meta data via CrossRef API? Default is TRUE. Currently, no useful other ways of identifying metadata are implemented.}

\item{bibtex}{Should reference be exported as bibtex? Works only with rcrossref = TRUE. Default is FALSE.}

\item{vars}{Which variables should be returned? Specify as character vectors. By default, all available variables are returned.}
}
\description{
Identifies meta information within txt documents.
}

