% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_wos_citations.R
\name{get_wos_citations}
\alias{get_wos_citations}
\title{Scrape citations from the web of knowledge.}
\usage{
get_wos_citations(wos.url, n.pages, folder)
}
\arguments{
\item{wos.url}{Follow the steps below to get the url.
on the first citation of the study.}

\item{n.pages}{Number of pages you would like to scrape.}

\item{folder}{Name the folder in which citations are saved.}
}
\value{
Returns a file 'citations.csv' with all the scraped citations of that particular study.
#'
}
\description{
Scrape citations from the web of knowledge.
}
\section{Steps to follow}{

1. Search for study in WOS.

2. Click on the "times cited" link on the right.

3. Click on the first result in the list.

4. Copy the link out of the adress line in the browser and insert that into the function.
}

