% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_citing_doc.R
\name{identify_citing_document}
\alias{identify_citing_document}
\title{Identifies meta information within documents.}
\usage{
identify_citing_document(folder, number = NULL)
}
\arguments{
\item{folder}{Name of folder within working directory in which the citing documents (.txt files) are located, e.g. "Beck 1995".}

\item{number}{Number of .txt files in folder the function should be applied to. Default is "all .txt files in folder".}
}
\value{
Stores list with meta information identified within citin documents (.txt files).
}
\description{
Identifies meta information within documents.
}

