% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identify_study_in_text.R
\name{identify_study}
\alias{identify_study}
\title{Identifies whether single documents (text files) really cite (or just contain) study.}
\usage{
identify_study(study.title)
}
\arguments{
\item{study.title}{Specify title of study.}
}
\value{
Returns a list of the extracted references/titles that correspond to the input title.
}
\description{
Identifies whether single documents (text files) really cite (or just contain) study.
}

