% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_study_doc.R
\name{identify_study_doc}
\alias{identify_study_doc}
\title{Identifies whether single documents (text files) really cite (or just contain) study.}
\usage{
identify_study_doc(study.title, folder, number = NULL)
}
\arguments{
\item{study.title}{Title of the study, e.g. "Measuring trust"}

\item{folder}{Name of folder within working directory in which the citing documents (.txt files) are located, e.g. "Beck 1995".}

\item{number}{Number of .txt files in folder the function should be applied to. Default is "all .txt files in folder".}
}
\value{
Returns a both a .csv and html files that contain a
list with the titles that were extracted across the citing documents.
}
\description{
Identifies whether single documents (text files) really cite (or just contain) study.
}
\examples{
\dontrun{
 setwd("C:/Users/paul/Google Drive/Research/2016_Quality_of_citations/data")
 folder <- "Acemoglu 2001"
 study.title <- "The colonial origins of comparative development: An empirical investigation"
 identify_study_doc(study.title, folder)
}
}

