% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_docs.R
\name{rename_docs}
\alias{rename_docs}
\title{Renames docs/txt files using the metadata.}
\usage{
rename_docs(folder)
}
\arguments{
\item{folder}{Name of folder within working directory in which the citing documents (.txt files) are located, e.g. "Beck 1995".}

\item{number}{Number of .txt files in folder the function should be applied to. Default is "all .txt files in folder".}
}
\value{
Returns and saves text files without running heads.
}
\description{
Renames docs/txt files using the metadata.
}

