% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_docs_names.R
\name{reset_doc_names}
\alias{reset_doc_names}
\title{Identifies meta information within txt documents.}
\usage{
reset_doc_names(folder = "docs", pdf = TRUE, txt = TRUE)
}
\arguments{
\item{folder}{Name of folder that contains the documents.}

\item{pdf}{Default TRUE. Set FALSE if only .txt files should be renamed.}

\item{txt}{Default TRUE. Set FALSE if only .pdf files should be renamed.}
}
\description{
Renames files in folder with rising index.
}
\examples{
\dontrun{
 setwd("C:/data")
 rename_docs(folder = "docs")
}
}
