% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study_data.R
\name{study_data}
\alias{study_data}
\title{Store data on the study of interest in ris file, elements of which can be used later on.}
\usage{
study_data(...)
}
\arguments{
\item{...}{Specify elements such as title etc.}
}
\value{
Should save study information in a ris file.
}
\description{
Store data on the study of interest in ris file, elements of which can be used later on.
}

