% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_analysis.R
\name{topic_analysis}
\alias{topic_analysis}
\title{Conducts topic analysis of citations of a given article}
\usage{
topic_analysis(file, article, output, K = 3:8, runs = 2, max.em.its = 10,
  net.max.em.its = 2, seed = 777)
}
\arguments{
\item{file}{'citation_data.csv' that contains data on the citation cases.}

\item{article}{Name of cited article; e.g. Fearon (2003)}

\item{output}{folder where topics object, figures, and tables generated by function will be stored}

\item{K}{number of topics to be computed. If more than one number, will run models with all and then
choose that with the highest semantic coherence.}

\item{runs}{total number of STM runs in the cast net stage}

\item{max.em.its}{maximum number of EM iterations in STM}

\item{net.max.em.its}{maximum number of EM iterations when casting the net with STM.}

\item{seed}{seed for random number generator in STM.}
}
\description{
Conducts topic analysis of citations of a given article
}
\examples{
\dontrun{
 file <- "~/Google Drive/2016_Quality_of_citations/data/Fearon 2003_citation_cases.csv"
 article <- "Fearon and Laitin (2003)"
 output <- "fearon_2003"
 topic_analysis(file, article, output)
} 
}

