% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_citations_cases.R
\name{extract_citation_cases}
\alias{extract_citation_cases}
\title{Extract citation cases.}
\usage{
extract_citation_cases(folder, authorname, studyyear, scope = NULL,
  number = NULL)
}
\arguments{
\item{folder}{Name of folder within working directory in which the citing documents (.txt files) are located, e.g. "Beck 1995".}

\item{authorname}{Enter the names of all authors separated by comma.}

\item{studyyear}{Enter the year when the study was published.}

\item{scope}{Number of sentences before/after citation case sentence to extract.}
}
\value{
Returns and saves data (.csv and .html) containing citation cases in current working directory.
}
\description{
Extract citation cases.
}

