% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_citations_cases.R
\name{extract_citation_cases}
\alias{extract_citation_cases}
\title{Extract citation cases from "cleaned_woref.txt" files.}
\usage{
extract_citation_cases()
}
\value{
Returns and saves data on citation cases (citation_data.csv).
}
\description{
Extract citation cases from "cleaned_woref.txt" files.
}

