% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{Extracts text from PDFs and stores in folder of choice.}
\usage{
extract_text(from, to, number = NULL, method = "pdftools",
  path.pdftotext = NULL)
}
\arguments{
\item{from}{Folder with .pdf files.}

\item{to}{Folder with .txt files.}

\item{number}{Number of PDFs you want to apply the function to. Default is "all PDFs in folder".}

\item{method}{Which method to use. Default is pdftotext.exe.}
}
\value{
Takes PDFs in folder "from" and extracts .txt files that are copied into folder "to".
}
\description{
Extracts text from PDFs and stores in folder of choice.
}
\examples{
\dontrun{
 folder <- "Acemoglu 2001"
 extract_text(folder)
}
}
