% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{Convert PDFs in /documents/ into text files which are stored in same folder.}
\usage{
extract_text()
}
\value{
Converts PDFs into text files that a stored in the "documents" folder.
}
\description{
Convert PDFs in /documents/ into text files which are stored in same folder.
}

