% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identify_study_in_text.R
\name{identify_study}
\alias{identify_study}
\title{Identifies whether single documents (text files) really cite (or just contain) study.}
\usage{
identify_study(study.title, folder, number = NULL)
}
\arguments{
\item{study.title}{Specify title of study.}

\item{folder}{The folder in which the citing documents are located.}
}
\value{
Returns a both a .csv file as well as a .html file that contain a
list with the titles that were extracted from all the citing documents.
}
\description{
Identifies whether single documents (text files) really cite (or just contain) study.
}

